+++ sdk/cpp/tests/test_path_api.cpp
@@ -668,3 +668,17 @@ TEST_CASE("iana_if_type_decode")
     REQUIRE(xml == ifc_payload);
 }
 
+TEST_CASE( "rpc_get_schema_no_decode" )
+{
+    ydk::path::Repository repo{};
+
+    ydk::path::NetconfSession session{repo,"127.0.0.1", "admin", "admin",  12022};
+    ydk::path::RootSchemaNode& schema = session.get_root_schema();
+
+    auto rpc = schema.create_rpc("ietf-netconf-monitoring:get-schema");
+    rpc->get_input_node().create_datanode("identifier", "ydktest-sanity-action");
+
+    auto reply = session.execute_netconf_operation(*rpc);
+    REQUIRE(!reply.empty());
+    //cout << reply << endl;
+}